Allow more than 10000 unique shuffles
Currently, the deck is shuffled based on a seed chosen at random between 0 and 10000. This means that there are only about 10000 possible deck shuffles, which is quite small; if you play 119 games, you have a greater than 50% chance of seeing the same shuffle twice at some point. This PR adds a lot more entropy to the shuffle seeds so that the chance of seeing the same shuffle seed twice is negligible.

Incidentally, there's a similar problem with the readable game IDs; based on the [unique-names-generator](https://www.npmjs.com/package/unique-names-generator) documentation, there are only about 1400 * 350 * 50 = 24.5 million possible IDs. Starting from an empty database, if 5829 games are created, one of them would likely error out due to a collision, and this will get more and more common as the game database grows indefinitely. This isn't as big a problem as the shuffle seed (if creating a game fails, you could just try again and likely succeed) but it could be a good idea to add more entropy to game IDs as well. (I think this could be done without degrading the UX by using more words in the IDs, but making the words much more common.)