Introduce Coded Hint markers.
One marker for color clues, and one for number clues.  It is a feature that can be enabled in User Preferences.   The intent is to make it clear what information a player has a quick glance.  With the corner hint marker, you have to hover or long-press to tell this.